* Fixing page_update() busy-wait that could deadlock the GUI thread when called reentrantly * Fixing shutdown-while-encoding bug where the worker would lose the shutdown intent after the current encode finished, requiring a forceful GUI kill instead of graceful shutdown
