+++ fastflix/data/languages.yaml
@@ -14701,3 +14701,18 @@ Rotation and flip will be set as display metadata for supported containers (MP4,
   ukr: Обертання та перевертання буде встановлено як метадані відображення для підтримуваних контейнерів (MP4, MOV, MKV).
   kor: 회전 및 뒤집기는 지원되는 컨테이너(MP4, MOV, MKV)의 표시 메타데이터로 설정됩니다.
   ron: Rotirea și întoarcerea vor fi setate ca metadate de afișare pentru containerele acceptate (MP4, MOV, MKV).
+Are you sure you want to delete all encoding history?:
+  eng: Are you sure you want to delete all encoding history?
+  deu: Sind Sie sicher, dass Sie den gesamten Kodierungsverlauf löschen wollen?
+  fra: Êtes-vous sûr de vouloir supprimer tout l'historique d'encodage ?
+  ita: Siete sicuri di voler cancellare tutta la cronologia delle codifiche?
+  spa: ¿Estás seguro de que quieres borrar todo el historial de codificación?
+  jpn: エンコード履歴をすべて削除してもよろしいですか？
+  rus: Вы уверены, что хотите удалить всю историю кодирования?
+  por: Tem a certeza de que pretende apagar todo o histórico de codificação?
+  swe: Är du säker på att du vill radera all kodningshistorik?
+  pol: Czy na pewno chcesz usunąć całą historię kodowania?
+  chs: 您确定要删除所有编码历史记录吗？
+  ukr: Ви впевнені, що хочете видалити всю історію кодування?
+  kor: 모든 인코딩 기록을 삭제하시겠습니까?
+  ron: Sunteți sigur că doriți să ștergeți tot istoricul de codare?
